added abstract class
