+++ finance_service/src/main/kotlin/com/internship/finance_service/service/finance/ReadFinanceOperationService.kt
@@ -13,6 +13,7 @@ import com.internship.finance_service.entity.WalletTransfer
 import com.internship.finance_service.enums.sort.OrderDirection
 import com.internship.finance_service.repo.PaymentRepo
 import com.internship.finance_service.repo.WalletTransferRepo
+import com.internship.finance_service.service.PageableObjectCreator
 import com.internship.finance_service.service.specification.PaymentSpecification
 import com.internship.finance_service.service.specification.WalletTransferSpecification
 import com.internship.finance_service.utils.FinanceValidationManager
@@ -31,13 +32,13 @@ class ReadFinanceOperationService @Autowired constructor(
     private var financeValidationManager: FinanceValidationManager,
     private var paymentMapper: PaymentMapper,
     private var walletTransferMapper: WalletTransferMapper
-) {
+): PageableObjectCreator(){
 
     @Transactional(readOnly = true)
     fun findAllPayments(filter: PaymentFilterRequest): PaymentPackageDto {
         val spec = PaymentSpecification.createFilterSpecification(filter)
 
-        val pageable = createPageableObject(filter.page, filter.size, filter.sortBy.toString(), filter.orderDirection)
+        val pageable = createPageableObjectByAllFields(filter.page, filter.size, filter.sortBy.toString(), filter.orderDirection)
         val page: Page<Payment> = paymentRepo.findAll(spec, pageable)
         val paymentsDto = page.content.map(paymentMapper::toDto)
 
@@ -54,7 +55,7 @@ class ReadFinanceOperationService @Autowired constructor(
     fun findAllWalletTransfers(filter: WalletTransferFilterRequest): WalletTransferPackageDto {
         val spec = WalletTransferSpecification.createFilterSpecification(filter)
 
-        val pageable = createPageableObject(filter.page, filter.size, filter.sortBy.toString(), filter.orderDirection)
+        val pageable = createPageableObjectByAllFields(filter.page, filter.size, filter.sortBy.toString(), filter.orderDirection)
         val page: Page<WalletTransfer> = walletTransferRepo.findAll(spec, pageable)
         val walletTransfersDto = page.content.map(walletTransferMapper::toDto)
 
@@ -75,20 +76,5 @@ class ReadFinanceOperationService @Autowired constructor(
     @Transactional(readOnly = true)
     fun getWalletTransferById(id: Long): WalletTransferDto {
         return walletTransferMapper.toDto(financeValidationManager.getWalletTransferOperationIfExists(id))
-    }
-
-    private fun createPageableObject(
-        page: Int,
-        size: Int,
-        sortByField: String,
-        orderDirection: OrderDirection
-    ): Pageable {
-        val sort = Sort.by(
-            Sort.Direction.fromString(
-                orderDirection.toString()
-            ),
-            sortByField
-        )
-        return PageRequest.of(page, size, sort)
     }
 }
